Refactor browser files. Add consistency. Fix some broken shortcuts
