+++ apps/brave/brave.talon
@@ -1,0 +1,4 @@
+app: brave
+-
+tag(): browser
+tag(): user.tabs
