+++ apps/chrome/chrome.py
@@ -10,11 +10,8 @@ and app.exe: chrome.exe
 """
 mod.apps.chrome = """
 os: mac
-and app.bundle: com.google.Chrome
-"""
-mod.apps.chrome = """
-os: mac
-and app.bundle: org.chromium.Chromium
+app.bundle: com.google.Chrome
+app.bundle: org.chromium.Chromium
 """
 mod.apps.chrome = """
 os: linux
@@ -25,13 +22,14 @@ mod.apps.chrome = """
 os: linux
 and app.name: Google-chrome
 """
+
 ctx.matches = r"""
 app: chrome
 """
 
 
 @mod.action_class
-class Actions:
+class UserActions:
     def chrome_mod(key: str):
         """Press the specified key with the correct modifier key for the OS"""
         if app.platform == "mac":
@@ -39,47 +37,13 @@ class Actions:
         else:
             actions.key(f"ctrl-{key}")
 
-
-@ctx.action_class("user")
-class user_actions:
-    def tab_jump(number: int):
-        if number < 9:
-            if app.platform == "mac":
-                actions.key(f"cmd-{number}")
-            else:
-                actions.key(f"ctrl-{number}")
-
-    def tab_final():
-        if app.platform == "mac":
-            actions.key("cmd-9")
-        else:
-            actions.key("ctrl-9")
-
     def tab_close_wrapper():
         actions.sleep("180ms")
         actions.app.tab_close()
 
-    def tab_duplicate():
-        actions.browser.focus_address()
-        actions.sleep("180ms")
-        possibly_edited_url = actions.edit.selected_text()
-        actions.key("esc")
-        actions.browser.focus_address()
-        actions.sleep("180ms")
-        url_address = actions.edit.selected_text()
-        if possibly_edited_url == url_address:
-            actions.key("alt-enter")
-        else:
-            actions.user.paste(possibly_edited_url)
-            actions.app.tab_open()
-            actions.user.paste(url_address)
-            actions.key("enter")
-
 
 @ctx.action_class("browser")
-class browser_actions:
-    def go(url: str):
-        actions.browser.focus_address()
-        actions.sleep("50ms")
-        actions.insert(url)
-        actions.key("enter")
+class BrowserActions:
+    def show_extensions():
+        actions.app.tab_open()
+        actions.browser.go("chrome://extensions")
