+++ apps/chrome/chrome.talon
@@ -1,1 +1,6 @@
+app: chrome
+-
+tag(): browser
+tag(): user.tabs
+
 profile switch: user.chrome_mod("shift-m")
