+++ apps/edge/edge.py
@@ -21,8 +21,6 @@ app: microsoft_edge
 
 @ctx.action_class("browser")
 class BrowserActions:
-    def go(url: str):
-        actions.browser.focus_address()
-        actions.sleep("50ms")
-        actions.insert(url)
-        actions.key("enter")
+    def show_extensions():
+        actions.app.tab_open()
+        actions.browser.go("edge://extensions")
