+++ apps/edge/edge.talon
@@ -1,0 +1,4 @@
+app: microsoft_edge
+-
+tag(): browser
+tag(): user.tabs
