+++ apps/firefox/firefox.talon
@@ -2,6 +2,7 @@ app: firefox
 -
 tag(): browser
 tag(): user.tabs
+
 tab search:
     browser.focus_address()
     insert("% ")
