+++ apps/safari/safari.talon
@@ -1,0 +1,4 @@
+app: safari
+-
+tag(): browser
+tag(): user.tabs
