+++ tags/browser/browser.py
@@ -1,6 +1,6 @@
 from urllib.parse import urlparse
 
-from talon import Context, actions
+from talon import Context, actions, app
 
 ctx = Context()
 ctx.matches = r"""
@@ -18,6 +18,25 @@ def is_url(url):
         return False
 
 
+@ctx.action_class("user")
+class UserActions:
+    def tab_duplicate():
+        actions.browser.focus_address()
+        actions.sleep("180ms")
+        possibly_edited_url = actions.edit.selected_text()
+        actions.key("esc")
+        actions.browser.focus_address()
+        actions.sleep("180ms")
+        url_address = actions.edit.selected_text()
+        if possibly_edited_url == url_address:
+            actions.key("alt-enter")
+        else:
+            actions.user.paste(possibly_edited_url)
+            actions.app.tab_open()
+            actions.user.paste(url_address)
+            actions.key("enter")
+
+
 @ctx.action_class("browser")
 class BrowserActions:
     def address():
@@ -28,3 +47,63 @@ class BrowserActions:
             for url in reversed(actions.win.title().split(" "))
         )
         return next((url for url in tokens if is_url(url)), "")
+
+    def bookmark():
+        actions.key("ctrl-d")
+
+    def bookmark_tabs():
+        actions.key("ctrl-shift-d")
+
+    def bookmarks():
+        actions.key("ctrl-shift-o")
+
+    def bookmarks_bar():
+        actions.key("ctrl-shift-b")
+
+    def focus_address():
+        actions.key("alt-d")
+
+    def focus_search():
+        actions.browser.focus_address()
+
+    def go_blank():
+        actions.key("ctrl-n")
+
+    def go(url: str):
+        actions.browser.focus_address()
+        actions.sleep("50ms")
+        actions.insert(url)
+        actions.key("enter")
+
+    def go_home():
+        actions.key("alt-home")
+
+    def go_back():
+        actions.key("alt-left")
+
+    def go_forward():
+        actions.key("alt-right")
+
+    def open_private_window():
+        actions.key("ctrl-shift-n")
+
+    def reload():
+        actions.key("ctrl-r")
+
+    def reload_hard():
+        actions.key("ctrl-shift-r")
+
+    def show_downloads():
+        actions.key("ctrl-j")
+
+    def show_clear_cache():
+        actions.key("ctrl-shift-backspace")
+
+    def show_history():
+        actions.key("ctrl-h")
+
+    def submit_form():
+        actions.key("enter")
+
+    def toggle_dev_tools():
+        actions.key("ctrl-shift-i")
