+++ tags/browser/browser.talon
@@ -29,44 +29,3 @@ show downloads: browser.show_downloads()
 show extensions: browser.show_extensions()
 show history: browser.show_history()
 show cache: browser.show_clear_cache()
-
-#todo - port to apps
-# navigating current page
-# help: key(?)
-# scroll tiny down: key(j)
-# scroll tiny up: key(k)
-# scroll left: key(h)
-# scroll right: key(l)
-# scroll (pop | spring): key(z H)
-# scroll push: key(z L)
-# scroll top: key(gg)
-# scroll (bottom | end): key(G)
-# (scroll half down | mini page): key(d)
-# scroll half up: key(u)
-# [open] link: key(f)
-# [open] link new: key(F)
-# copy link: key(y f)
-# copy (address | url): key(escape y y)
-# (refresh | reload): browser.reload()
-# view source: key(g s)
-# insert mode: key(i)
-# next frame: key(g f)
-# main frame: key(g F)
-# navigating to new pages
-# (open | go) (url | history): key(o)
-# (open | go) (url | history) new: key(O)
-# (open | go) bookmark: key(b)
-# (open | go) bookmark new: key(B)
-# using find
-# find mode: key(/)
-# next match: key(n)
-# previous match: key(N)
-# navigating history
-# history back: key(H)
-# history forward: key(L)
-# manipulating tabs
-# last visited: key(^)
-# dupe tab: key(y t)
-# restore: key(X)
-# search tabs: key(T)
-# move to window: key(W)
